refactoring, better debug output, disassembler

+++ timer.cpp
@@ -6,9 +6,9 @@ void Timer::reset() {
 	m_timestamp = std::chrono::high_resolution_clock::now();
 }
 
-long Timer::framedeltaMicroseconds() {
+uint32_t Timer::framedeltaMicroseconds() {
 	auto diff = std::chrono::high_resolution_clock::now() - m_timestamp;
-	long microSeconds = std::chrono::duration_cast<std::chrono::microseconds>(diff).count();
+	uint32_t microSeconds = std::chrono::duration_cast<std::chrono::microseconds>(diff).count();
 	reset();
 	return microSeconds;
 }
