+++ timer.h
@@ -7,7 +7,7 @@ class Timer {
 public:
 	Timer();
 	void reset();
-	long framedeltaMicroseconds();
+	uint32_t framedeltaMicroseconds();
 private:
 	std::chrono::high_resolution_clock::time_point m_timestamp;
 };
